requirements.txt
alembic==1.8.1
attrs==22.1.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
Flask==2.2.0
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
Mako==1.2.1
MarkupSafe==2.1.1
pkgutil-resolve-name==1.3.10
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.1
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.2.1
zipp==3.8.1
